+++ CHANGELOG.md
@@ -9,12 +9,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 
 - Created a set of instructions for how to set up a local development environment for the toolkit. This is available in the LOCAL_DEVELOPMENT.md file.
+- Added image export button when creating a build.
 
 ### Changed
 
 - Changed item cards to be more uniform. No more multi-column layout on desktop, too much empty space.
 - Updated the images for Bloat King boss and Little Gorge aberration. (ConRaven)
 - Removed DESCRIPTION_TAGS from autocomplete search boxes as it is no longer necessary.
+- Split BUILD SUGGESTIONS button in the builder into two buttons: Armor Calculator and Item Suggestions. This should make it more clear utilities are available.
 
 ### Fixed